corrected spelling of sylph in main window
