+++ package.json
@@ -11,7 +11,6 @@
     "out:win": "rollup -c && tsc && electron-builder build --win --publish never",
     "publish:win": "rollup -c && tsc && electron-builder build --win --publish always"
   },
-
   "devDependencies": {
     "@rollup/plugin-commonjs": "^17.0.0",
     "@rollup/plugin-node-resolve": "^11.0.0",
@@ -43,6 +42,7 @@
     "sirv-cli": "^1.0.12",
     "svelte-ace": "^1.0.21",
     "svelte-select": "^4.3.1",
+    "svelte-spa-router": "^3.2.0",
     "svelte-split-pane": "^0.1.2"
   }
 }
